mapping for jeanselector added
